encoder: add VBR mode driven by per-subband masking
Introduces variable-bitrate Layer I encoding alongside the existing
CBR path. The per-subband masking model (`psy::SubbandMask`) drives a
two-phase allocator: Phase 1 grants upgrades until every band's
quantisation noise is below its mask threshold; Phase 2 spends any
remaining headroom up to a quality-scaled fraction of a rolling
per-frame target derived from `vbr_target_kbps`. Per-frame
`bitrate_index` floats over the standard ladder (32..=448 kbps); the
long-term average converges on the requested target via the running
surplus controller.

Mask model is the lightweight per-subband variant (energy + one-tap
inter-band spreading + ATH boost above 12 kHz) — not a full ISO
§C.1 psymodel-1 implementation; that's noted as a follow-up.

Selected via `vbr_target_kbps` and/or `vbr_quality` (0..=9) on
`CodecParameters::options`. CBR remains the default.

Tests: +7 unit (psy module) + 8 end-to-end (encoder_vbr) covering
steady-tone target tracking, transient-vs-steady per-frame
variation, multi-second average within target ± expected variance,
distinct bitrate-slot usage, own-decoder roundtrip and ffmpeg
cross-decode interop. Total crate test count 31 -> 46.

+++ tests/encode_roundtrip.rs
@@ -89,7 +89,7 @@ fn decode_all(data: &[u8]) -> (Vec<i16>, u32, u16) {
         if sr == 0 {
             if let Ok(h) = FrameHeader::parse(fr) {
                 sr = h.sample_rate;
-                ch = h.mode.channel_count() as u16;
+                ch = h.mode.channel_count();
             }
         }
         let pkt = Packet {
